Crowbar 0.51 - 2017-11-19_0.51
Changelog:

=== Unpack ===

    FIXED: Unpacking from a selected folder that is direct child of <root> folder in the tree view crashes Crowbar.
    FIXED: Unpacking from a "<Found>" tree (search results) can sometimes crash Crowbar.
    Technical details: Crashes when a found folder name, such as "k_lab", is part of an unfound folder name, such as "k_lab2".
    FIXED: Clicking Subfolder "Use Default" button changes the Compile tab's Subfolder option instead of the Unpack tab's.
    FIXED: "Use in Preview" button sets the first game setup in Preview instead of not changing it.
    FIXED: After unpacking, the list in the bottom combobox of MDL file names is not sorted alphabetically.
    FIXED: Write to a log option does not write a file.
    FIXED: Entering "..\unpacked" for Output Subfolder changes the textbox to and places output in: "<path_of_crowbar.exe>\unpacked".
    FIXED: Unpacking from a non-multi-chunk version 2 VPK is incorrect.

=== Decompile ===

    FIXED: MDL v10 does not decompile $sequence options: pivot, node, transition, rtransition.
    FIXED: MDL v10 file with a modelname that does not end with ".mdl" writes a $modelname in the QC file that does not re-compile.
    Example stored modelname (without quotes) that shows problem: "C:\Program Files\Steam\SteamApps\longishusername\half-life\SvenC"
    FIXED: MDL v2531 (Vampire the Masquerade Bloodlines) - Many physics meshes are decompiled incorrectly.
    FIXED: MDL v2531 - VertexType1 and VertexType2 meshes are decompiled incorrectly.
    FIXED: MDL v37 physics and animation SMD files are not created.
    FIXED: MDL v37 animations are rotated 90 degrees.
    FIXED: MDL v49 eyelid flexes sometimes use the wrong frames from the VTA file. Reported here.
    FIXED: The QC $poseparameter "loop" option is not decompiled for all MDL versions that use $poseparameter.
    FIXED: Options groupbox does not disable when decompiling.
    FIXED: Entering "..\decompiled" for Output Subfolder changes the textbox to and places output in: "<path_of_crowbar.exe>\decompiled".
    ADDED: MDL v48 and v49 - Near-equivalent QC flex expressions for DMX-only flex expressions are now written to QC file.

=== Compile ===

    FIXED: MDL v10 compiling can seemingly freeze the Compile tab if compiling to a folder that is rooted to a large folder.
    FIXED: Entering "..\compiled" for Output Subfolder changes the textbox to and places output in: "<path_of_crowbar.exe>\compiled".
    FIXED: If compile fails, then using Goto on the selected log file does not go to the correct folder.
    FIXED: Incorrectly deletes empty folders that are part of the path to the temp folders created by Crowbar.

=== Options ===

    UPDATED: Changed the defaults for Windows Explorer Auto-Open so that all checkboxes are checkmarked.

+++ Crowbar/Core/Viewer/Viewer.vb
@@ -105,17 +105,21 @@ Public Class Viewer
 			'RemoveHandler Me.theHlmvAppProcess.Exited, AddressOf HlmvApp_Exited
 
 			Try
-				If Not Me.theHlmvAppProcess.CloseMainWindow() Then
+				If Not Me.theHlmvAppProcess.HasExited AndAlso Not Me.theHlmvAppProcess.CloseMainWindow() Then
 					Me.theHlmvAppProcess.Kill()
 				End If
 			Catch ex As Exception
 				Dim debug As Integer = 4242
 			Finally
-				Me.theHlmvAppProcess.Close()
-				'NOTE: This raises an exception when the background thread has already completed its work.
-				'If calledFromBackgroundThread Then
-				'	Me.UpdateProgressStop("Model viewer closed.")
-				'End If
+				'NOTE: Due to threading, Me.theHlmvAppProcess might be Nothing at this point.
+				If Me.theHlmvAppProcess IsNot Nothing Then
+					Me.theHlmvAppProcess.Close()
+					'NOTE: This raises an exception when the background thread has already completed its work.
+					'If calledFromBackgroundThread Then
+					'	Me.UpdateProgressStop("Model viewer closed.")
+					'End If
+					Me.theHlmvAppProcess = Nothing
+				End If
 			End Try
 		End If
 	End Sub
